blur done (need debug)
-main.cpp
The current kernel is 1D 13-samples kernel.
In the first pass, it will get the color, blur and depth information.
In the second pass, it will blur the image from first pass in the
vertical direction.
In the third pass, it will blur the image from second pass in the
horizontal direction.